+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
+### Added
+
+- Field sync task: Empty only option for extract pattern
+
 ## [0.40.6] - 2024-04-30
 
 ### Changed